GOB: Do not narrow Expression value type from int32 to int16 too early
This fixes a bug in Adibou2/Sciences where Adibou kept talking without
pauses, because some game variable storing the time in milliseconds
overflowed int16.

+++ engines/gob/inter_playtoons.cpp
@@ -355,7 +355,7 @@ void Inter_Playtoons::oPlaytoons_loadMultObject() {
 }
 
 void Inter_Playtoons::oPlaytoons_getObjAnimSize() {
-	int16 objIndex;
+	int32 objIndex;
 	uint16 readVar[4];
 	uint16 types[4];
 	Mult::Mult_AnimData animData;
